--- src/main/java/com/example/whiplash/article/document/Article.java
@@ -1,22 +1,0 @@
-package com.example.whiplash.article.document;
-
-import jakarta.persistence.GeneratedValue;
-import lombok.*;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.time.LocalDateTime;
-
-@Document(collection = "articles")
-@Getter @Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-public class Article {
-    @Id @GeneratedValue
-    private String id;
-    private String title;
-    private String press;
-    private LocalDateTime publishedAt;
-    private Category category;
-}
