+++ src/main/java/com/example/whiplash/article/domain/document/SummarizedArticle.java
@@ -1,12 +1,8 @@
-package com.example.whiplash.article.document;
+package com.example.whiplash.article.domain.document;
 
 import com.example.whiplash.domain.entity.history.email.SummaryLevel;
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
 import lombok.*;
-import org.hibernate.annotations.Index;
 import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
