+++ src/main/java/com/example/whiplash/article/domain/entity/UserArticleAssignment.java
@@ -1,4 +1,4 @@
-package com.example.whiplash.article.entity;
+package com.example.whiplash.article.domain.entity;
 
 import com.example.whiplash.domain.entity.history.email.EmailSendStatus;
 import com.example.whiplash.user.domain.User;
