+++ src/main/java/com/example/whiplash/article/repository/ArticleRepository.java
@@ -1,6 +1,6 @@
 package com.example.whiplash.article.repository;
 
-import com.example.whiplash.article.document.Article;
+import com.example.whiplash.article.domain.document.Article;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.time.LocalDateTime;
