+++ src/main/java/com/example/whiplash/delivery/assignment/ArticleAssignerV1.java
@@ -1,8 +1,8 @@
 package com.example.whiplash.delivery.assignment;
 
-import com.example.whiplash.article.document.Category;
-import com.example.whiplash.article.document.SummarizedArticle;
-import com.example.whiplash.article.entity.UserArticleAssignment;
+import com.example.whiplash.article.domain.document.Category;
+import com.example.whiplash.article.domain.document.SummarizedArticle;
+import com.example.whiplash.article.domain.entity.UserArticleAssignment;
 import com.example.whiplash.article.repository.SummarizedArticleRepository;
 import com.example.whiplash.article.repository.UserArticleAssignmentRepository;
 import com.example.whiplash.domain.entity.history.email.EmailSendStatus;
